[TRITON] Adapt Model Benchmarking Scripts to the New `bench_mha.py` CLI (#2673)
* Fix wrong data type argument in `bench_mha.py` sink generation

* Adapt `bench_attn_models.py` to new `bench_mha.py` CLI

* Adapt `bench_models.py` to new `bench_mha.py` CLI

* Adapt `bench_attn_models.py` to new `bench_mla_decode.py` CLI

+++ op_tests/op_benchmarks/triton/bench_mha.py
@@ -598,7 +598,9 @@ def run_benchmark(run: BenchRun):
             requires_grad=requires_grad,
         )
         sink = (
-            torch.randn((HQ,), device=device, dtype=dtype, requires_grad=requires_grad)
+            torch.randn(
+                (HQ,), device=device, dtype=torch_dtype, requires_grad=requires_grad
+            )
             if run.sink
             else None
         )
